+++ Pandora/Core.hs
@@ -1,5 +1,4 @@
 module Pandora.Core (module Exports) where
 
 import Pandora.Core.Impliable as Exports
-import Pandora.Core.Appliable as Exports
 import Pandora.Core.Functor as Exports